+++ terraform/modules/aks/variables.tf
@@ -1,0 +1,2 @@
+variable "resource_group_name" { type = string }
+variable "location" { type = string }