+++ terraform/modules/auth0/outputs.tf
@@ -1,0 +1,2 @@
+output "app_client_id" { value = auth0_client.app_spa.client_id }
+output "vault_client_id" { value = auth0_client.vault_client.client_id }