+++ terraform/modules/auth0/versions.tf
@@ -1,0 +1,6 @@
+terraform {
+  required_version = ">= 1.5.0"
+  required_providers {
+    auth0 = { source = "auth0/auth0", version = "~> 1.0" }
+  }
+}