+++ terraform/modules/compute/outputs.tf
@@ -1,0 +1,1 @@
+output "public_ip_address" { value = azurerm_linux_virtual_machine.vm.public_ip_address }