+++ terraform/modules/registry/outputs.tf
@@ -1,0 +1,2 @@
+output "login_server" { value = azurerm_container_registry.acr.login_server }
+output "name" { value = azurerm_container_registry.acr.name }