+++ terraform/modules/registry/variables.tf
@@ -1,0 +1,4 @@
+variable "resource_group_name" { type = string }
+variable "location" { type = string }
+variable "name_suffix" { type = string }
+variable "aks_kubelet_object_id" { type = string }